Update README.md
In order for people to use this library it is very important that the documentation is readable and that it shows that it is taken care of very well. Therefore I suggest you have somebody check the text for spelling errors and other mistakes. I improved the documentation somewhat to get you started.